refactor: rename namespace from 'util' to 'utils' for consistency
