+++ includes/Utils.hpp
@@ -18,10 +18,10 @@
 
 #include "Utils.tpp"
 
-namespace util
+namespace utils
 {
 
 std::string strtrim(const std::string& str);
 std::string quoted(const std::string& str, bool double_quotes = false);
 
-} // namespace util
+} // namespace utils
